'Final_Code'

+++ controllers/post_controller.js
@@ -88,8 +88,9 @@ try{
         // CHANGE :: delete the associated likes for the post and all its comments' likes too
         await Like.deleteMany({likeable: post, onModel: 'Post'});
         await Like.deleteMany({_id: {$in: post.comments}});
-
+        if(post.avatar){
         await fs.unlinkSync(path.join(__dirname,'..', post.avatar));
+        }
         post.remove();
         await Comment.deleteMany({post:req.params.id});
